pyproject.toml
[tool.poetry]
name = "marigold-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"
PyMySQL = "^1.0.2"
djangorestframework-simplejwt = "^5.0.0"
cryptography = "^35.0.0"
django-extensions = "^3.1.5"
django-cors-headers = "^3.10.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
